Moved all PKINIT structures into the one file
